requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
cryptography==41.0.7
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0